move copy test data into common testig tools
