+++ runner/common_test.go
@@ -58,6 +58,29 @@ type File struct {
 	isDir bool
 }
 
+// copyTestData will make a tempdir and then
+// "cp -r" a subdirectory under testdata there
+// returns the directory (which can now be used as config.Config.Home) and modified safely
+func copyTestData(subdir string) (string, error) {
+	tmpdir, err := ioutil.TempDir("", "pocket-runner-test")
+	if err != nil {
+		return "", errors.Wrap(err, "create temp dir")
+	}
+
+	src := filepath.Join("testdata", subdir)
+
+	options := Options{
+		Recursive: true,
+		Atomic:    true,
+	}
+	err = Copy(src, tmpdir, options)
+	if err != nil {
+		os.RemoveAll(tmpdir)
+		return "", errors.Wrap(err, "copying files")
+	}
+	return tmpdir, nil
+}
+
 // NewFile creates a new File.
 func NewFile(path string) *File {
 	return &File{Path: path}
